Using Handlebars for templates
Moved a lot of HTML-code out of the JavaScript files into Handlebars
(http://handlebarsjs.com/) script tags.
It’s planned to make Abfrage3 available in multiple languages with PHP
and therefore every “content” should be in HTML-files.
There is still a lot of HTML-code in JS-files which will be moved later.

Added some comments.