Fixed inconsistency with HTML-encoding
Now the server takes not encoded strings (e.g. “a & b”) converts them
(htmlspecialchars(): “a &amp; b”) and stores them in the data base.

When the client requests them later the strings will be unescaped again
(htmlspecialchars_decode(): “a & b”) and the client has to escape them
(if necessary, e.g. when showing them in the DOM).

Since Handlebars escapes anyway this is not a big deal client side.